+++ libcloud/compute/drivers/gce.py
@@ -519,6 +519,33 @@ class GCENodeImage(NodeImage):
                                               deprecated, obsolete, deleted)
 
 
+class GCESubnetwork(UuidMixin):
+    """A GCE Subnetwork object class."""
+    def __init__(self, id, name, cidr, network, region, driver, extra=None):
+        self.id = str(id)
+        self.name = name
+        self.cidr = cidr
+        self.network = network
+        self.region = region
+        self.driver = driver
+        self.extra = extra
+        UuidMixin.__init__(self)
+
+    def destroy(self):
+        """
+        Destroy this subnetwork
+
+        :return: True if successful
+        :rtype:  ``bool``
+        """
+        return self.driver.ex_destroy_subnetwork(self)
+
+    def __repr__(self):
+        return '<GCESubnetwork id="%s" name="%s" region="%s" network="%s" ' \
+               'cidr="%s">' % (self.id, self.name, self.region.name,
+                               self.network.name, self.cidr)
+
+
 class GCENetwork(UuidMixin):
     """A GCE Network object class."""
     def __init__(self, id, name, cidr, driver, extra=None):
@@ -527,6 +554,11 @@ class GCENetwork(UuidMixin):
         self.cidr = cidr
         self.driver = driver
         self.extra = extra
+        self.mode = 'legacy'
+        self.subnetworks = []
+        if 'mode' in extra and extra['mode'] != 'legacy':
+            self.mode = extra['mode']
+            self.subnetworks = extra['subnetworks']
         UuidMixin.__init__(self)
 
     def destroy(self):
@@ -539,8 +571,8 @@ class GCENetwork(UuidMixin):
         return self.driver.ex_destroy_network(network=self)
 
     def __repr__(self):
-        return '<GCENetwork id="%s" name="%s" cidr="%s">' % (
-            self.id, self.name, self.cidr)
+        return '<GCENetwork id="%s" name="%s" cidr="%s" mode="%s">' % (
+            self.id, self.name, self.cidr, self.mode)
 
 
 class GCERoute(UuidMixin):
@@ -1587,6 +1619,43 @@ class GCENodeDriver(NodeDriver):
                        response.get('items', [])]
         return list_routes
 
+    def ex_list_subnetworks(self, region=None):
+        """
+        Return the list of subnetworks.
+
+        :keyword  region: Region for the subnetwork. Specify 'all' to return
+                          the aggregated list of subnetworks.
+        :type     region: ``str`` or :class:`GCERegion`
+
+        :return: A list of subnetwork objects.
+        :rtype: ``list`` of :class:`GCESubnetwork`
+        """
+        region = self._set_region(region)
+        if region is None:
+            request = '/aggregated/subnetworks'
+        else:
+            request = '/regions/%s/subnetworks' % (region.name)
+
+        list_subnetworks = []
+        response = self.connection.request(request, method='GET').object
+
+        if 'items' in response:
+            if region is None:
+                for v in response['items'].values():
+                    for i in v.get('subnetworks', []):
+                        try:
+                            list_subnetworks.append(self._to_subnetwork(i))
+                        except ResourceNotFoundError:
+                            pass
+            else:
+                for i in response['items']:
+                    try:
+                        list_subnetworks.append(self._to_subnetwork(i))
+                    except ResourceNotFoundError:
+                        pass
+
+        return list_subnetworks
+
     def ex_list_networks(self):
         """
         Return the list of networks.
@@ -2252,28 +2321,111 @@ class GCENodeDriver(NodeDriver):
 
         return self.ex_get_route(name)
 
-    def ex_create_network(self, name, cidr, description=None):
+    def ex_create_subnetwork(self, name, cidr=None, network=None, region=None,
+                             description=None):
         """
-        Create a network.
+        Create a subnetwork.
 
-        :param  name: Name of network to be created
+        :param  name: Name of subnetwork to be created
         :type   name: ``str``
 
         :param  cidr: Address range of network in CIDR format.
         :type   cidr: ``str``
 
+        :param  network: The network name or object this subnet belongs to.
+        :type   network: ``str`` or :class:`GCENetwork`
+
+        :param  region: The region the subnetwork belongs to.
+        :type   region: ``str`` or :class:`GCERegion`
+
         :param  description: Custom description for the network.
         :type   description: ``str`` or ``None``
+
+        :return:  Subnetwork object
+        :rtype:   :class:`GCESubnetwork`
+        """
+        if not cidr:
+            raise ValueError("Must provide an IP network in CIDR notation.")
+
+        if not network:
+            raise ValueError("Must provide a network for the subnetwork.")
+        else:
+            if isinstance(network, GCENetwork):
+                network_url = network.extra['selfLink']
+            else:
+                if network.startswith('https://'):
+                    network_url = network
+                else:
+                    network_obj = self.ex_get_network(network)
+                    network_url = network_obj.extra['selfLink']
+
+        if not region:
+            raise ValueError("Must provide a region for the subnetwork.")
+        else:
+            if isinstance(region, GCERegion):
+                region_url = region.extra['selfLink']
+            else:
+                if region.startswith('https://'):
+                    region_url = region
+                else:
+                    region_obj = self.ex_get_region(region)
+                    region_url = region_obj.extra['selfLink']
+
+        subnet_data = {}
+        subnet_data['name'] = name
+        subnet_data['description'] = description
+        subnet_data['ipCidrRange'] = cidr
+        subnet_data['network'] = network_url
+        subnet_data['region'] = region_url
+        region_name = region_url.split('/')[-1]
+
+        request = '/regions/%s/subnetworks' % (region_name)
+        self.connection.async_request(request, method='POST',
+                                      data=subnet_data)
+
+        return self.ex_get_subnetwork(name, region_name)
+
+    def ex_create_network(self, name, cidr, description=None, mode="legacy"):
+        """
+        Create a network. In November 2015, Google introduced Subnetworks and
+        suggests using networks with 'auto' generated subnetworks. See, the
+        `subnet docs <https://cloud.google.com/compute/docs/subnetworks>`_ for
+        more details. Note that libcloud follows the usability pattern from
+        the Cloud SDK (e.g. 'gcloud compute' command-line utility) and uses
+        'mode' to specify 'auto', 'custom', or 'legacy'.
+
+        :param  name: Name of network to be created
+        :type   name: ``str``
+
+        :param  cidr: Address range of network in CIDR format.
+        :type   cidr: ``str`` or ``None``
+
+        :param  description: Custom description for the network.
+        :type   description: ``str`` or ``None``
+
+        :param  mode: Create a 'auto', 'custom', or 'legacy' network.
+        :type   mode: ``str``
 
         :return:  Network object
         :rtype:   :class:`GCENetwork`
         """
         network_data = {}
         network_data['name'] = name
-        network_data['IPv4Range'] = cidr
         network_data['description'] = description
+        if mode.lower() not in ['auto', 'custom', 'legacy']:
+            raise ValueError("Invalid network mode: '%s'. Must be 'auto', "
+                             "'custom', or 'legacy'." % mode)
+        if cidr and mode in ['auto', 'custom']:
+            raise ValueError("Can only specify IPv4Range with 'legacy' mode.")
 
         request = '/global/networks'
+
+        if mode == 'legacy':
+            if not cidr:
+                raise ValueError("Must specify IPv4Range with 'legacy' mode.")
+            network_data['IPv4Range'] = cidr
+        else:
+            network_data['autoCreateSubnetworks'] = (mode.lower() == 'auto')
 
         self.connection.async_request(request, method='POST',
                                       data=network_data)
@@ -2281,7 +2433,8 @@ class GCENodeDriver(NodeDriver):
         return self.ex_get_network(name)
 
     def create_node(self, name, size, image, location=None,
-                    ex_network='default', ex_tags=None, ex_metadata=None,
+                    ex_network='default', ex_subnetwork=None,
+                    ex_tags=None, ex_metadata=None,
                     ex_boot_disk=None, use_existing_disk=True,
                     external_ip='ephemeral', ex_disk_type='pd-standard',
                     ex_disk_auto_delete=True, ex_service_accounts=None,
@@ -2308,6 +2461,9 @@ class GCENodeDriver(NodeDriver):
 
         :keyword  ex_network: The network to associate with the node.
         :type     ex_network: ``str`` or :class:`GCENetwork`
+
+        :keyword  ex_subnetwork: The subnetwork to associate with the node.
+        :type     ex_subnetwork: ``str`` or :class:`GCESubnetwork`
 
         :keyword  ex_tags: A list of tags to associate with the node.
         :type     ex_tags: ``list`` of ``str`` or ``None``
@@ -2431,6 +2587,12 @@ class GCENodeDriver(NodeDriver):
             size = self.ex_get_size(size, location)
         if not hasattr(ex_network, 'name'):
             ex_network = self.ex_get_network(ex_network)
+        if ex_subnetwork:
+            if not hasattr(ex_subnetwork, 'name'):
+                ex_subnetwork = \
+                    self.ex_get_subnetwork(ex_subnetwork,
+                                           region=self._get_region_from_zone(
+                                               location))
         if ex_image_family:
             image = self.ex_get_image_from_family(ex_image_family)
         if image and not hasattr(image, 'name'):
@@ -2468,7 +2630,8 @@ class GCENodeDriver(NodeDriver):
                                                    ex_nic_gce_struct,
                                                    ex_on_host_maintenance,
                                                    ex_automatic_restart,
-                                                   ex_preemptible)
+                                                   ex_preemptible,
+                                                   ex_subnetwork)
         self.connection.async_request(request, method='POST', data=node_data)
         return self.ex_get_node(name, location.name)
 
@@ -4254,6 +4417,91 @@ class GCENodeDriver(NodeDriver):
         response = self.connection.request(request, method='GET').object
         return self._to_route(response)
 
+    def ex_destroy_subnetwork(self, name, region=None):
+        """
+        Delete a Subnetwork object based on name and region.
+
+        :param  name: The name, URL or object of the subnetwork
+        :type   name: ``str`` or :class:`GCESubnetwork`
+
+        :param  name: The region object, name, or URL of the subnetwork
+        :type   name: ``str`` or :class:`GCERegion` or ``None``
+
+        :return:  True if successful
+        :rtype:   ``bool``
+        """
+        region_name = None
+        subnet_name = None
+        if region:
+            if isinstance(region, GCERegion):
+                region_name = region.name
+            else:
+                if region.startswith('https://'):
+                    region_name = region.split('/')[-1]
+                else:
+                    region_name = region
+        if isinstance(name, GCESubnetwork):
+            subnet_name = name.name
+            if not region_name:
+                region_name = name.region.name
+        else:
+            if name.startswith('https://'):
+                url_parts = self._get_components_from_path(name)
+                subnet_name = url_parts['name']
+                if not region_name:
+                    region_name = url_parts['region']
+            else:
+                subnet_name = name
+
+        if not region_name:
+            region = self._set_region(region)
+            if not region:
+                raise ("Could not determine region for subnetwork.")
+            else:
+                region_name = region.name
+
+        request = '/regions/%s/subnetworks/%s' % (region_name, subnet_name)
+        self.connection.request(request, method='DELETE').object
+        return True
+
+    def ex_get_subnetwork(self, name, region=None):
+        """
+        Return a Subnetwork object based on name and region.
+
+        :param  name: The name or URL of the subnetwork
+        :type   name: ``str``
+
+        :param  name: The region of the subnetwork
+        :type   name: ``str`` or :class:`GCERegion` or ``None``
+
+        :return:  A Subnetwork object
+        :rtype:   :class:`GCESubnetwork`
+        """
+        region_name = None
+        if name.startswith('https://'):
+            parts = self._get_components_from_path(name)
+            name = parts['name']
+            region_name = parts['region']
+        else:
+            if isinstance(region, GCERegion):
+                region_name = region.name
+            elif isinstance(region, str):
+                if region.startswith('https://'):
+                    region_name = region.split('/')[-1]
+                else:
+                    region_name = region
+
+        if not region_name:
+            region = self._set_region(region)
+            if not region:
+                raise ("Could not determine region for subnetwork.")
+            else:
+                region_name = region.name
+
+        request = '/regions/%s/subnetworks/%s' % (region_name, name)
+        response = self.connection.request(request, method='GET').object
+        return self._to_subnetwork(response)
+
     def ex_get_network(self, name):
         """
         Return a Network object based on a network name.
@@ -4719,7 +4967,7 @@ class GCENodeDriver(NodeDriver):
                          ex_disks_gce_struct=None, ex_nic_gce_struct=None,
                          ex_on_host_maintenance=None,
                          ex_automatic_restart=None,
-                         ex_preemptible=None):
+                         ex_preemptible=None, ex_subnetwork=None):
         """
         Returns a request and body to create a new node.  This is a helper
         method to support both :class:`create_node` and
@@ -4830,6 +5078,9 @@ class GCENodeDriver(NodeDriver):
                                          not be preemptible)
         :type     ex_preemptible: ``bool`` or ``None``
 
+        :param  ex_subnetwork: The network to associate with the node.
+        :type   ex_subnetwork: :class:`GCESubnetwork`
+
         :return:  A tuple containing a request string and a node_data dict.
         :rtype:   ``tuple`` of ``str`` and ``dict``
         """
@@ -4934,9 +5185,12 @@ class GCENodeDriver(NodeDriver):
                                      "'ex_nic_gce_struct'. Use one or the "
                                      "other.")
 
+        ni = []
         if network:
             ni = [{'kind': 'compute#instanceNetworkInterface',
                    'network': network.extra['selfLink']}]
+            if ex_subnetwork:
+                ni[0]['subnetwork'] = ex_subnetwork.extra['selfLink']
             if external_ip:
                 access_configs = [{'name': 'External NAT',
                                    'type': 'ONE_TO_ONE_NAT'}]
@@ -5336,6 +5590,33 @@ class GCENodeDriver(NodeDriver):
                                  protocol=forwarding_rule.get('IPProtocol'),
                                  targetpool=target, driver=self, extra=extra)
 
+    def _to_subnetwork(self, subnetwork):
+        """
+        Return a Subnetwork object from the JSON-response dictionary.
+
+        :param  subnetwork: The dictionary describing the subnetwork.
+        :type   subnetwork: ``dict``
+
+        :return: Subnetwork object
+        :rtype: :class:`GCESubnetwork`
+        """
+        extra = {}
+
+        extra['creationTimestamp'] = subnetwork.get('creationTimestamp')
+        extra['description'] = subnetwork.get('description')
+        extra['gatewayAddress'] = subnetwork.get('gatewayAddress')
+        extra['ipCidrRange'] = subnetwork.get('ipCidrRange')
+        extra['network'] = subnetwork.get('network')
+        extra['region'] = subnetwork.get('region')
+        extra['selfLink'] = subnetwork.get('selfLink')
+        network = self._get_object_by_kind(subnetwork.get('network'))
+        region = self._get_object_by_kind(subnetwork.get('region'))
+
+        return GCESubnetwork(id=subnetwork['id'], name=subnetwork['name'],
+                             cidr=subnetwork.get('ipCidrRange'),
+                             network=network, region=region,
+                             driver=self, extra=extra)
+
     def _to_network(self, network):
         """
         Return a Network object from the JSON-response dictionary.
@@ -5349,9 +5630,23 @@ class GCENodeDriver(NodeDriver):
         extra = {}
 
         extra['selfLink'] = network.get('selfLink')
-        extra['gatewayIPv4'] = network.get('gatewayIPv4')
         extra['description'] = network.get('description')
         extra['creationTimestamp'] = network.get('creationTimestamp')
+        # 'legacy'
+        extra['gatewayIPv4'] = network.get('gatewayIPv4')
+        extra['IPv4Range'] = network.get('IPv4Range')
+        # 'auto' or 'custom'
+        extra['autoCreateSubnetworks'] = network.get('autoCreateSubnetworks')
+        extra['subnetworks'] = network.get('subnetworks')
+
+        # match Cloud SDK 'gcloud'
+        if 'autoCreateSubnetworks' in network:
+            if network['autoCreateSubnetworks']:
+                extra['mode'] = 'auto'
+            else:
+                extra['mode'] = 'custom'
+        else:
+            extra['mode'] = 'legacy'
 
         return GCENetwork(id=network['id'], name=network['name'],
                           cidr=network.get('IPv4Range'),
@@ -5529,8 +5824,12 @@ class GCENodeDriver(NodeDriver):
         extra['guestCpus'] = machine_type.get('guestCpus')
         extra['creationTimestamp'] = machine_type.get('creationTimestamp')
         try:
+            orig_api_name = self.api_name
+            self.api_name = "%s_%s" % (self.api_name,
+                                       extra['zone'].name.split("-")[0])
             price = self._get_size_price(size_id=machine_type['name'])
-        except KeyError:
+            self.api_name = orig_api_name
+        except:
             price = None
 
         return GCENodeSize(id=machine_type['id'], name=machine_type['name'],