+++ libcloud/test/compute/test_gce.py
@@ -1,4 +1,4 @@
-# Licensed to the Apache Software Foundation (ASF) under one or more
+# License to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
 # this work for additional information regarding copyright ownership.
 # The ASF licenses this file to You under the Apache License, Version 2.0
@@ -25,9 +25,9 @@ from libcloud.compute.drivers.gce import (GCENodeDriver, API_VERSION,
                                           GCEAddress, GCEBackendService,
                                           GCEFirewall, GCEForwardingRule,
                                           GCEHealthCheck, GCENetwork,
-                                          GCENodeImage, GCERoute,
+                                          GCENodeImage, GCERoute, GCERegion,
                                           GCETargetHttpProxy, GCEUrlMap,
-                                          GCEZone)
+                                          GCEZone, GCESubnetwork)
 from libcloud.common.google import (GoogleBaseAuthConnection,
                                     ResourceNotFoundError, ResourceExistsError,
                                     InvalidRequestError, GoogleBaseError)
@@ -128,6 +128,7 @@ class GCENodeDriverTest(GoogleTestCase, TestCaseMixin):
                         d.ex_list_forwarding_rules,
                         d.ex_list_healthchecks,
                         d.ex_list_networks,
+                        d.ex_list_subnetworks,
                         d.ex_list_project_images,
                         d.ex_list_regions,
                         d.ex_list_routes,
@@ -233,10 +234,72 @@ class GCENodeDriverTest(GoogleTestCase, TestCaseMixin):
         self.assertEqual(len(routes), 3)
         self.assertTrue('lcdemoroute' in [route.name for route in routes])
 
+    def test_ex_list_subnetworks(self):
+        subnetworks = self.driver.ex_list_subnetworks()
+        self.assertEqual(len(subnetworks), 1)
+        self.assertEqual(subnetworks[0].name, 'cf-972cf02e6ad49112')
+        self.assertEqual(subnetworks[0].cidr, '10.128.0.0/20')
+        subnetworks = self.driver.ex_list_subnetworks('all')
+        self.assertEqual(len(subnetworks), 4)
+
+    def test_ex_create_subnetwork(self):
+        name = 'cf-972cf02e6ad49112'
+        cidr = '10.128.0.0/20'
+        network_name = 'cf'
+        network = self.driver.ex_get_network(network_name)
+        region_name = 'us-central1'
+        region = self.driver.ex_get_region(region_name)
+        # test by network/region name
+        subnet = self.driver.ex_create_subnetwork(name, cidr, network_name, region_name)
+        self.assertTrue(isinstance(subnet, GCESubnetwork))
+        self.assertTrue(isinstance(subnet.region, GCERegion))
+        self.assertTrue(isinstance(subnet.network, GCENetwork))
+        self.assertEqual(subnet.name, name)
+        self.assertEqual(subnet.cidr, cidr)
+        # test by network/region object
+        subnet = self.driver.ex_create_subnetwork(name, cidr, network, region)
+        self.assertTrue(isinstance(subnet, GCESubnetwork))
+        self.assertTrue(isinstance(subnet.region, GCERegion))
+        self.assertTrue(isinstance(subnet.network, GCENetwork))
+        self.assertEqual(subnet.name, name)
+        self.assertEqual(subnet.cidr, cidr)
+
+    def test_ex_destroy_subnetwork(self):
+        name = 'cf-972cf02e6ad49112'
+        region_name = 'us-central1'
+        region = self.driver.ex_get_region(region_name)
+        # delete with no region
+        self.assertTrue(self.driver.ex_destroy_subnetwork(name))
+        # delete with region name
+        self.assertTrue(self.driver.ex_destroy_subnetwork(name, region_name))
+        # delete with region object
+        self.assertTrue(self.driver.ex_destroy_subnetwork(name, region))
+
+    def test_ex_get_subnetwork(self):
+        name = 'cf-972cf02e6ad49112'
+        region_name = 'us-central1'
+        region = self.driver.ex_get_region(region_name)
+        # fetch by no region
+        subnetwork = self.driver.ex_get_subnetwork(name)
+        self.assertEqual(subnetwork.name, name)
+        # fetch by region name
+        subnetwork = self.driver.ex_get_subnetwork(name, region_name)
+        self.assertEqual(subnetwork.name, name)
+        # fetch by region object
+        subnetwork = self.driver.ex_get_subnetwork(name, region)
+        self.assertEqual(subnetwork.name, name)
+
     def test_ex_list_networks(self):
         networks = self.driver.ex_list_networks()
         self.assertEqual(len(networks), 3)
-        self.assertEqual(networks[0].name, 'default')
+        self.assertEqual(networks[0].name, 'cf')
+        self.assertEqual(networks[0].mode, 'auto')
+        self.assertEqual(len(networks[0].subnetworks), 4)
+        self.assertEqual(networks[1].name, 'custom')
+        self.assertEqual(networks[1].mode, 'custom')
+        self.assertEqual(len(networks[1].subnetworks), 1)
+        self.assertEqual(networks[2].name, 'default')
+        self.assertEqual(networks[2].mode, 'legacy')
 
     def test_list_nodes(self):
         nodes = self.driver.list_nodes()
@@ -479,6 +542,16 @@ class GCENodeDriverTest(GoogleTestCase, TestCaseMixin):
         self.assertEqual(network.name, network_name)
         self.assertEqual(network.cidr, cidr)
 
+    def test_ex_create_network_bad_options(self):
+        network_name = 'lcnetwork'
+        cidr = '10.11.0.0/16'
+        self.assertRaises(ValueError, self.driver.ex_create_network,
+                          network_name, cidr, mode='auto')
+        self.assertRaises(ValueError, self.driver.ex_create_network,
+                          network_name, cidr, mode='foobar')
+        self.assertRaises(ValueError, self.driver.ex_create_network,
+                          network_name, None, mode='legacy')
+
     def test_ex_set_machine_type_notstopped(self):
         # get running node, change machine type
         zone = 'us-central1-a'
@@ -581,6 +654,46 @@ class GCENodeDriverTest(GoogleTestCase, TestCaseMixin):
         self.assertRaises(ValueError, self.driver.create_node, node_name,
                           size, image, location=zone, ex_network=network,
                           ex_nic_gce_struct=ex_nic_gce_struct)
+
+    def test_create_node_subnetwork_opts(self):
+        node_name = 'sn-node-name'
+        size = self.driver.ex_get_size('n1-standard-1')
+        image = self.driver.ex_get_image('debian-7')
+        zone = self.driver.ex_get_zone('us-central1-a')
+        network = self.driver.ex_get_network('custom-network')
+        subnetwork = self.driver.ex_get_subnetwork('cf-972cf02e6ad49112')
+
+        ex_nic_gce_struct = [
+            {
+                "network": "global/networks/custom-network",
+                "subnetwork": "projects/project_name/regions/us-central1/subnetworks/cf-972cf02e6ad49112",
+                "accessConfigs": [
+                    {
+                        "name": "External NAT",
+                        "type": "ONE_TO_ONE_NAT"
+                    }
+                ]
+            }
+        ]
+        # Test using just the network and subnetwork
+        node = self.driver.create_node(node_name, size, image, location=zone,
+                                       ex_network=network,
+                                       ex_subnetwork=subnetwork)
+        self.assertEqual(node.extra['networkInterfaces'][0]["name"], 'nic0')
+        self.assertEqual(node.extra['networkInterfaces'][0]["subnetwork"].split('/')[-1], 'cf-972cf02e6ad49112')
+
+        # Test using just the struct
+        node = self.driver.create_node(node_name, size, image, location=zone,
+                                       ex_nic_gce_struct=ex_nic_gce_struct)
+        self.assertEqual(node.extra['networkInterfaces'][0]["name"], 'nic0')
+        self.assertEqual(node.extra['networkInterfaces'][0]["subnetwork"].split('/')[-1], 'cf-972cf02e6ad49112')
+
+        # Test using subnetwork selfLink
+        node = self.driver.create_node(node_name, size, image, location=zone,
+                                       ex_network=network,
+                                       ex_subnetwork=subnetwork.extra['selfLink'])
+        self.assertEqual(node.extra['networkInterfaces'][0]["name"], 'nic0')
+        self.assertEqual(node.extra['networkInterfaces'][0]["subnetwork"].split('/')[-1], 'cf-972cf02e6ad49112')
 
     def test_create_node_disk_opts(self):
         node_name = 'node-name'
@@ -1573,6 +1686,10 @@ class GCEMockHttp(MockHttpTestCase):
             body = self.fixtures.load('setCommonInstanceMetadata_post.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
 
+    def _aggregated_subnetworks(self, method, url, body, headers):
+        body = self.fixtures.load('aggregated_subnetworks.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
     def _aggregated_addresses(self, method, url, body, headers):
         body = self.fixtures.load('aggregated_addresses.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
@@ -1731,6 +1848,14 @@ class GCEMockHttp(MockHttpTestCase):
             body = self.fixtures.load('global_networks_post.json')
         else:
             body = self.fixtures.load('global_networks.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _global_networks_custom_network(self, method, url, body, headers):
+        body = self.fixtures.load('global_networks_custom_network.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _global_networks_cf(self, method, url, body, headers):
+        body = self.fixtures.load('global_networks_cf.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
 
     def _global_networks_default(self, method, url, body, headers):
@@ -1949,6 +2074,11 @@ class GCEMockHttp(MockHttpTestCase):
             body = self.fixtures.load('global_urlMaps_web_map.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
 
+    def _regions_us_central1_subnetworks_cf_972cf02e6ad49112(self, method, url, body, headers):
+        body = self.fixtures.load(
+            'regions_us-central1_subnetworks_cf_972cf02e6ad49112.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
     def _regions_us_central1_operations_operation_regions_us_central1_addresses_lcaddress_delete(
             self, method, url, body, headers):
         body = self.fixtures.load(
@@ -1965,6 +2095,12 @@ class GCEMockHttp(MockHttpTestCase):
             self, method, url, body, headers):
         body = self.fixtures.load(
             'operations_operation_regions_us-central1_addresses_post.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _regions_us_central1_operations_operation_regions_us_central1_subnetworks_post(
+            self, method, url, body, headers):
+        body = self.fixtures.load(
+            'operations_operation_regions_us-central1_subnetworks_post.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
 
     def _regions_us_central1_operations_operation_regions_us_central1_forwardingRules_post(
@@ -2232,6 +2368,30 @@ class GCEMockHttp(MockHttpTestCase):
             body = self.fixtures.load('global_addresses.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
 
+    def _regions_europe_west1(self, method, url, body, headers):
+        body = self.fixtures.load('regions_europe-west1.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _regions_asia_east1(self, method, url, body, headers):
+        body = self.fixtures.load('regions_asia-east1.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _regions_us_central1(self, method, url, body, headers):
+        body = self.fixtures.load('regions_us-central1.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _regions_us_east1(self, method, url, body, headers):
+        body = self.fixtures.load('regions_us-east1.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _regions_us_central1_subnetworks(self, method, url, body, headers):
+        if method == 'POST':
+            body = self.fixtures.load(
+                'regions_us-central1_subnetworks_post.json')
+        else:
+            body = self.fixtures.load('regions_us-central1_subnetworks.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
     def _regions_us_central1_addresses(self, method, url, body, headers):
         if method == 'POST':
             body = self.fixtures.load(
@@ -2384,6 +2544,14 @@ class GCEMockHttp(MockHttpTestCase):
         body = self.fixtures.load('zones_asia-east1-a.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
 
+    def _zones_asia_east1_b(self, method, url, body, headers):
+        body = self.fixtures.load('zones_asia-east1-b.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _zones_us_east1_b(self, method, url, body, headers):
+        body = self.fixtures.load('zones_us-east1-b.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
     def _zones_us_central1_a_diskTypes(self, method, url, body, headers):
         body = self.fixtures.load('zones_us-central1-a_diskTypes.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
@@ -2506,6 +2674,11 @@ class GCEMockHttp(MockHttpTestCase):
                 'zones_us-central1-a_instances_post.json')
         else:
             body = self.fixtures.load('zones_us-central1-a_instances.json')
+        return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
+
+    def _zones_us_central1_a_instances_sn_node_name(self, method, url, body,
+                                                    headers):
+        body = self.fixtures.load('zones_us-central1-a_instances_sn-node-name.json')
         return (httplib.OK, body, self.json_hdr, httplib.responses[httplib.OK])
 
     def _zones_us_central1_a_instances_node_name(self, method, url, body,